+++ src/main/java/com/cursomc/models/PagamentoComBoleto.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 
 @Entity
 public class PagamentoComBoleto extends Pagamento{
-	private static final long serialVersionUID = 1L;
+	private static final Long serialVersionUID = 1L;
 	
 	@JsonFormat(pattern="dd/MM/yyyy")
 	private Date dataVencimento;
@@ -21,7 +21,7 @@ public class PagamentoComBoleto extends Pagamento{
 	
 	
 	
-	public PagamentoComBoleto(Long id, EstadoPagamento estado, Pedido pedido, Date dataVencimento,Date dataPagamento ) {
+	public PagamentoComBoleto(Integer id, EstadoPagamento estado, Pedido pedido, Date dataVencimento,Date dataPagamento ) {
 		super(id, estado, pedido);
 		this.dataPagamento = dataPagamento;
 		this.dataVencimento = dataVencimento;
