+++ src/main/java/com/cursomc/models/PagamentoComCartao.java
@@ -6,19 +6,19 @@ import com.cursomc.models.enums.EstadoPagamento;
 
 @Entity
 public class PagamentoComCartao extends Pagamento{
-	private static final long serialVersionUID = 1L;
+	private static final Long serialVersionUID = 1L;
 
 	private Integer numeroDeParcelas;
 	
 	
 	
 	
-	public PagamentoComCartao(Long id, EstadoPagamento estado, Pedido pedido,Integer numeroDeParcelas) {
+	public PagamentoComCartao(Integer id, EstadoPagamento estado, Pedido pedido,Integer numeroDeParcelas) {
 		super(id, estado, pedido);
 		this.numeroDeParcelas = numeroDeParcelas;
 	}
 
-	public PagamentoComCartao(Long id) {
+	public PagamentoComCartao(Integer id) {
 	}
