+++ src/main/java/com/cursomc/repositories/CidadeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cursomc.models.Cidade;
 
 @Repository
-public interface CidadeRepository extends JpaRepository<Cidade, Long> {
+public interface CidadeRepository extends JpaRepository<Cidade, Integer> {
 	
 }
