+++ src/main/java/com/cursomc/repositories/EstadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cursomc.models.Estado;
 
 @Repository
-public interface EstadoRepository extends JpaRepository<Estado, Long> {
+public interface EstadoRepository extends JpaRepository<Estado, Integer> {
 	
 }
