+++ src/main/java/com/cursomc/services/AbstractService.java
@@ -2,6 +2,6 @@ package com.cursomc.services;
 
 public interface AbstractService {
 	
-	public <T> T findById(Long id);
+	public <T> T findById(Integer id);
 	
 }
